+++ lib/policy/config/config.go
@@ -25,6 +25,7 @@ var (
 	ErrInvalidUserAgentRegex             = errors.New("config.Bot: invalid user agent regex")
 	ErrInvalidPathRegex                  = errors.New("config.Bot: invalid path regex")
 	ErrInvalidHeadersRegex               = errors.New("config.Bot: invalid headers regex")
+	ErrInvalidDomainRegex                = errors.New("config.Bot: invalid domain regex")
 	ErrInvalidCIDR                       = errors.New("config.Bot: invalid CIDR")
 	ErrRegexEndsWithNewline              = errors.New("config.Bot: regular expression ends with newline (try >- instead of > in yaml)")
 	ErrInvalidImportStatement            = errors.New("config.ImportStatement: invalid source file")
@@ -59,6 +60,7 @@ type BotConfig struct {
 	UserAgentRegex *string           `json:"user_agent_regex,omitempty" yaml:"user_agent_regex,omitempty"`
 	PathRegex      *string           `json:"path_regex,omitempty" yaml:"path_regex,omitempty"`
 	HeadersRegex   map[string]string `json:"headers_regex,omitempty" yaml:"headers_regex,omitempty"`
+	DomainRegex    *string           `json:"domain_regex,omitempty" yaml:"domain_regex,omitempty"`
 	Expression     *ExpressionOrList `json:"expression,omitempty" yaml:"expression,omitempty"`
 	Challenge      *ChallengeRules   `json:"challenge,omitempty" yaml:"challenge,omitempty"`
 	Weight         *Weight           `json:"weight,omitempty" yaml:"weight,omitempty"`
@@ -80,6 +82,7 @@ func (b BotConfig) Zero() bool {
 		b.Action != "",
 		len(b.RemoteAddr) != 0,
 		b.Challenge != nil,
+		b.DomainRegex != nil,
 		b.GeoIP != nil,
 		b.ASNs != nil,
 	} {
@@ -156,7 +159,11 @@ func (b *BotConfig) Valid() error {
 			}
 		}
 	}
-
+	if b.DomainRegex != nil {
+		if _, err := regexp.Compile(*b.DomainRegex); err != nil {
+			errs = append(errs, ErrInvalidDomainRegex, err)
+		}
+	}
 	if b.Expression != nil {
 		if err := b.Expression.Valid(); err != nil {
 			errs = append(errs, err)