+++ lib/policy/policy.go
@@ -8,6 +8,7 @@ import (
 	"log/slog"
 	"sync/atomic"
 
+	"github.com/TecharoHQ/anubis/internal/fcrdns"
 	"github.com/TecharoHQ/anubis/internal/thoth"
 	"github.com/TecharoHQ/anubis/lib/policy/checker"
 	"github.com/TecharoHQ/anubis/lib/policy/config"
@@ -54,6 +55,7 @@ func ParseConfig(ctx context.Context, fin io.Reader, fname string, defaultDiffic
 	var validationErrs []error
 
 	tc, hasThothClient := thoth.FromContext(ctx)
+	fcrdns, hasFCrDNS := fcrdns.FromContext(ctx)
 
 	result := NewParsedConfig(c)
 	result.DefaultDifficulty = defaultDifficulty
@@ -108,8 +110,9 @@ func ParseConfig(ctx context.Context, fin io.Reader, fname string, defaultDiffic
 		}
 
 		if b.Expression != nil {
-			c, err := NewCELChecker(b.Expression)
-			if err != nil {
+			if !hasFCrDNS {
+				validationErrs = append(validationErrs, fmt.Errorf("while processing rule %s: no FCrDNS client in the context. This is a bug", b.Name))
+			} else if c, err := NewCELChecker(b.Expression, fcrdns); err != nil {
 				validationErrs = append(validationErrs, fmt.Errorf("while processing rule %s expressions: %w", b.Name, err))
 			} else {
 				cl = append(cl, c)
@@ -132,6 +135,16 @@ func ParseConfig(ctx context.Context, fin io.Reader, fname string, defaultDiffic
 			}
 
 			cl = append(cl, tc.GeoIPCheckerFor(b.GeoIP.Countries))
+		}
+
+		if b.DomainRegex != nil {
+			if !hasFCrDNS {
+				validationErrs = append(validationErrs, fmt.Errorf("while processing rule %s: no FCrDNS client in the context. This is a bug", b.Name))
+			} else if c, err := NewFCrDNSChecker(fcrdns, *b.DomainRegex); err != nil {
+				validationErrs = append(validationErrs, fmt.Errorf("while processing rule %s domain regex: %w", b.Name, err))
+			} else {
+				cl = append(cl, c)
+			}
 		}
 
 		if b.Challenge == nil {