+++ lib/policy/policy_test.go
@@ -7,11 +7,14 @@ import (
 
 	"github.com/TecharoHQ/anubis"
 	"github.com/TecharoHQ/anubis/data"
+	"github.com/TecharoHQ/anubis/internal/fcrdns"
 	"github.com/TecharoHQ/anubis/internal/thoth/thothmock"
 )
 
 func TestDefaultPolicyMustParse(t *testing.T) {
 	ctx := thothmock.WithMockThoth(t)
+	fdns := fcrdns.NewFCrDNS()
+	ctx = fcrdns.With(ctx, fdns)
 
 	fin, err := data.BotPolicies.Open("botPolicies.json")
 	if err != nil {
@@ -25,7 +28,6 @@ func TestDefaultPolicyMustParse(t *testing.T) {
 }
 
 func TestGoodConfigs(t *testing.T) {
-
 	finfos, err := os.ReadDir("config/testdata/good")
 	if err != nil {
 		t.Fatal(err)
@@ -42,6 +44,8 @@ func TestGoodConfigs(t *testing.T) {
 				defer fin.Close()
 
 				ctx := thothmock.WithMockThoth(t)
+				fdns := fcrdns.NewFCrDNS()
+				ctx = fcrdns.With(ctx, fdns)
 				if _, err := ParseConfig(ctx, fin, fin.Name(), anubis.DefaultDifficulty); err != nil {
 					t.Fatal(err)
 				}
@@ -54,7 +58,10 @@ func TestGoodConfigs(t *testing.T) {
 				}
 				defer fin.Close()
 
-				if _, err := ParseConfig(t.Context(), fin, fin.Name(), anubis.DefaultDifficulty); err != nil {
+				fdns := fcrdns.NewFCrDNS()
+				ctx := fcrdns.With(t.Context(), fdns)
+
+				if _, err := ParseConfig(ctx, fin, fin.Name(), anubis.DefaultDifficulty); err != nil {
 					t.Fatal(err)
 				}
 			})
@@ -63,8 +70,6 @@ func TestGoodConfigs(t *testing.T) {
 }
 
 func TestBadConfigs(t *testing.T) {
-	ctx := thothmock.WithMockThoth(t)
-
 	finfos, err := os.ReadDir("config/testdata/bad")
 	if err != nil {
 		t.Fatal(err)
@@ -78,6 +83,10 @@ func TestBadConfigs(t *testing.T) {
 				t.Fatal(err)
 			}
 			defer fin.Close()
+
+			ctx := thothmock.WithMockThoth(t)
+			fdns := fcrdns.NewFCrDNS()
+			ctx = fcrdns.With(ctx, fdns)
 
 			if _, err := ParseConfig(ctx, fin, fin.Name(), anubis.DefaultDifficulty); err == nil {
 				t.Fatal(err)